Full stack app

+++ hello-world/src/main/java/com/devops/helloworld/repositery/HelloWorldRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HelloWorldRepository extends JpaRepository<HelloWorld,Long> {
+public interface HelloWorldRepository extends JpaRepository<HelloWorld,Integer> {
 }
